+++ docs/news.rst
@@ -11,7 +11,7 @@ New features:
 
 Removed features:
 
-* Dropped the deprecated ``wait_for_reactor`` and unlimited timeouts on ``EventualResult.wait()``.
+* Dropped the deprecated APIs ``@wait_for_reactor``, ``@in_reactor``, ``DeferredResult``, and unlimited timeouts on ``EventualResult.wait()``.
 * Dropped support for Python 2.7 and 3.5.
 
 1.12.0